move event viewsets
